Change prism_id alphabet to exclude chances of generating words. (#2)
Generated IDs could previously include accidental crude or profane words due to the presence of vowels in the Base32 alphabet. This change replaces the Crockford alphabet with a custom one that excludes vowels to reduce the chance of generating inappropriate substrings. As a tradeoff, we allow visually similar characters (e.g., 0/O, 1/I).